integrasi

+++ src/component/Navbar.js
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router";
+import Swal from "sweetalert2";
 
 function Navbar({ toggleSidebar }) {
   const navigate = useNavigate();
@@ -10,8 +11,28 @@ function Navbar({ toggleSidebar }) {
   // };
   const token = localStorage.getItem("token");
   const logout = () => {
-    localStorage.clear();
-    navigate("/");
+    Swal.fire({
+      title: "Logout",
+      text: "Are you sure you want to logout?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, logout",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();
+        Swal.fire({
+          title: "Berhasil Logout",
+          text: "Anda berhasil keluar!",
+          icon: "success",
+          timer: 2000,
+          showConfirmButton: false,
+        }).then(() => {
+          navigate("/");
+        });
+      }
+    });
   };
 
   return (
@@ -25,11 +46,13 @@ function Navbar({ toggleSidebar }) {
             height="40"
             className=""
             viewBox="0 0 184.000000 183.000000"
-            preserveAspectRatio="xMidYMid meet">
+            preserveAspectRatio="xMidYMid meet"
+          >
             <g
               transform="translate(0.000000,183.000000) scale(0.100000,-0.100000)"
               fill="#776b5d"
-              stroke="none">
+              stroke="none"
+            >
               <path
                 d="M477 1684 c-4 -4 -7 -16 -7 -26 0 -13 7 -18 26 -18 21 0 25 4 22 22
 -3 23 -27 35 -41 22z"
@@ -107,13 +130,15 @@ m-320 -60 l0 -55 -55 0 -55 0 0 55 0 55 55 0 55 0 0 -55z m70 -69 c6 -8 10
               <div className="sm:flex sm:gap-4">
                 <a
                   className="block rounded-md bg-[#776b5d] px-5 py-2.5 text-sm font-medium text-white transition-all duration-200 shadow-md hover:bg-[#776b5d]/75 active:bg-gray-100 border border-[#776b5d] active:text-[#776b5d]"
-                  href="/sign-in">
+                  href="/sign-in"
+                >
                   Login
                 </a>
 
                 <a
                   className="hidden rounded-md bg-gray-100 px-5 py-2.5 text-sm font-medium text-[#776b5d] transition-all duration-200 hover:bg-[#776b5d] hover:shadow-md hover:text-white active:bg-gray-100 border border-transparent active:border-[#776b5d] active:text-[#776b5d] sm:block"
-                  href="/sign-up">
+                  href="/sign-up"
+                >
                   Register
                 </a>
               </div>
@@ -121,7 +146,8 @@ m-320 -60 l0 -55 -55 0 -55 0 0 55 0 55 55 0 55 0 0 -55z m70 -69 c6 -8 10
               <div className="sm:flex sm:gap-4">
                 <button
                   onClick={logout}
-                  className="block rounded-md bg-[#776b5d] px-5 py-2.5 text-sm font-medium text-white transition-all duration-200 shadow-md hover:bg-[#776b5d]/75 active:bg-gray-100 border border-[#776b5d] active:text-[#776b5d]">
+                  className="block rounded-md bg-[#776b5d] px-5 py-2.5 text-sm font-medium text-white transition-all duration-200 shadow-md hover:bg-[#776b5d]/75 active:bg-gray-100 border border-[#776b5d] active:text-[#776b5d]"
+                >
                   Logout
                 </button>
               </div>
@@ -129,7 +155,8 @@ m-320 -60 l0 -55 -55 0 -55 0 0 55 0 55 55 0 55 0 0 -55z m70 -69 c6 -8 10
 
             <button
               className="block rounded bg-gray-100 p-2.5 text-gray-600 transition hover:text-gray-600/75 md:hidden"
-              onClick={toggleSidebar}>
+              onClick={toggleSidebar}
+            >
               <span className="sr-only">Toggle menu</span>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
@@ -137,7 +164,8 @@ m-320 -60 l0 -55 -55 0 -55 0 0 55 0 55 55 0 55 0 0 -55z m70 -69 c6 -8 10
                 fill="none"
                 viewBox="0 0 24 24"
                 stroke="currentColor"
-                strokeWidth="2">
+                strokeWidth="2"
+              >
                 <path
                   strokeLinecap="round"
                   strokeLinejoin="round"
